Refactored Redux code
